moves errors to it's own file
